% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/getLRT.R
\name{getLRT}
\alias{getLRT}
\alias{getLRT_cpp}
\title{Compute the LRT using Dutilleul (1999)'s Algorithm}
\usage{
getLRT_cpp(Y, p, q, eps, Mean)

getLRT(Y, p, q, eps = 1e-10, Mean = rep(1, nrow(Y)) \%x\%
  matrix(colMeans(Y), nrow = 1))
}
\arguments{
\item{Y}{A r-by-(\code{p}*\code{q}) matrix where r is the sample size. See `Details' in \code{\link{Min_PermTest}} for its form.}

\item{p, q}{The number of rows and columns for matrix-variate data.}

\item{eps}{Tolerance constant for convergence. Default is 1e-10.}

\item{Mean}{The mean matrix of \code{Y}. Default is the sample mean vector for each row.}
}
\value{
A list with components:
\item{lrt}{The LRT statistic.}
\item{U}{Row-wise covariance (p-by-p).}
\item{V}{Column-wise covariance (q-by-q).}
}
\description{
This function implements Dutilleul (1999)'s Algorithm to obtain the likelihood ratio test statistic under normality. The null hypothesis is the separabiliy of covariance matrix, i.e. \eqn{\Sigma}=\code{U} \eqn{\otimes} \code{V}, and the alternative is not the null.

\code{\link{getLRT}} is an wrapper function for \code{\link{getLRT_cpp}} to set default parameters.
}
\examples{
U <- diag(rep(1,2))
V <- diag(rep(1,3))
set.seed(6)
Y <- mvtnorm::rmvnorm(50, sigma = U \%x\% V)
getLRT(Y, 2, 3)$lrt
}
\references{
Dutilleul, P. (1999). The mle algorithm for the matrix normal distribution. \emph{Journal of Statistical Computation and Simulation}, \strong{64(2)}, 105-123.
}
\seealso{
\code{\link{Min_PermTest}}
}
