% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUVPermMat.R
\name{getUVPermMat}
\alias{getUVPermMat}
\title{Generate a permutation matrix to ensure transition from row-wise permutation to column-wise.}
\usage{
getUVPermMat(p, q)
}
\arguments{
\item{p, q}{The number of rows and columns for matrix-variate data.}
}
\value{
A binary matrix for permutation purpose.
}
\description{
This function generates a permutation matrix P that rearranges a vector of \eqn{[(Y_{11}, ..., Y_{1q}, Y_{21}, ..., Y_{2q}, ..., Y_{pq})]}.
Here, row and column are inherited from the matrix form of it. If P is left-multiplied to the vector, the reordered vector is \eqn{[(Y_{11}, ..., Y_{p1}, Y_{12}, ..., Y_{p2}, ..., Y_{pq})]}.
}
\examples{
Y <- matrix(1:6, 2, 3, byrow = TRUE) # Imagine this is a matrix-variate data
vecY <- matrix(t(Y), ncol = 1) # identical with c(1,2,3,4,5,6)
P <- getUVPermMat(2, 3)
P \%*\% vecY # identical with c(1,4,2,5,3,6)
}
