% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermY.R
\name{PermY}
\alias{PermY}
\title{Perform random permutation of each row in \code{Y}}
\usage{
PermY(Y)
}
\arguments{
\item{Y}{A matrix. See `Details' for its form.}
}
\value{
The permuted matrix.
}
\description{
This function permutes two vectors in every rows depending on randomness. Therefore, permutation will be conducted as many as the number of rows.
}
\details{
For \code{Y}, each row is combined of two vectors to be permuted. In other words, the first half and the other half of a row would be exchanged with probability 1/2.
}
\examples{
set.seed(6)
matrix(rep(c(1,1,1,2,2,2), 6), nrow = 6, byrow = TRUE)
PermY(matrix(rep(c(1,1,1,2,2,2), 6), nrow = 6, byrow = TRUE))
}
