% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Row_PermTest.R
\name{Row_PermTest}
\alias{Row_PermTest}
\title{Perform the row-wise permutation test.}
\usage{
Row_PermTest(Y, p, q, N_perm = 2000, center = F, scale = T,
  n_cluster = 1)
}
\description{
The central function to test permutation tests.
}
\details{
See \code{\link{PermTest}} for arguments.
}
